feat: protect admin routes

+++ src/app/api/booking/[id]/route.ts
@@ -1,8 +1,18 @@
 import { NextRequest, NextResponse } from "next/server";
 
 import prisma from "@/lib/db";
+import { auth } from "@/auth";
 
 export async function PATCH(request: NextRequest, { params }: any) {
+  const session = await auth();
+
+  if (!session?.user) {
+    return NextResponse.json(
+      { error: true, message: "Unauthenticated" },
+      { status: 401 }
+    );
+  }
+
   const id = await params.id;
 
   const data = await request.json();
@@ -27,7 +37,17 @@ export async function PATCH(request: NextRequest, { params }: any) {
 }
 
 export async function DELETE(_: NextRequest, { params }: any) {
+  const session = await auth();
+
+  if (!session?.user) {
+    return NextResponse.json(
+      { error: true, message: "Unauthenticated" },
+      { status: 401 }
+    );
+  }
+
   const id = await params.id;
+
   try {
     await prisma.booking.delete({
       where: {
